Dynamic Modal
This is a piece of code that creates an simple container in an dynamic modal.
Problem: Imagine a list (10 items). In each item, we have a button. Pressing this button, we need open a container or modal. Could we repeat the modal inside the list? Oh, yeah. But, this is not the best semantic solution.
Solution: We create a single container. Pressing the button, we get the data from button and send to container. That´s is the dynamic modal.

Enjoy!